+++ src/api/mod.rs
@@ -1,4 +1,5 @@
 mod image_type;
+mod send_request;
 
 pub mod configuration;
 pub mod delete_image;
@@ -8,3 +9,4 @@ pub mod upload_image;
 
 pub use configuration::ImgurHandle;
 pub use image_type::*;
+pub use send_request::send_api_request;